Created a standard method for decreasing health. Added HP & status color codes.
decHealth method gets a value and a cause for the decrease, updates the HP stat and color code if needed. If the player reaches a non-positive HP value then it is killed, using the cause specified.
There is still work to do to change every hp decrease to use this method. Definition of constant color codes is also needed.