Fixed some issues with world map controls and UI.
There was an issue with the key control for world map if the user closed via the close button. Having a close button does not fit the game style anyway, so I removed it. I also got rid of the 'toggleWorldMap' function which was already available in the generic 'toggleWindow' function with some tweaks. Also did some minor style modifications.